+++ resolwe_bio/processes/import_data/bam_scseq.py
@@ -15,14 +15,14 @@ class ImportScBam(Process):
     slug = "upload-bam-scseq-indexed"
     name = "Single cell BAM file and index"
     process_type = "data:alignment:bam:scseq"
-    version = "1.2.1"
+    version = "1.3.0"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     entity = {"type": "sample"}
     requirements = {
         "expression-engine": "jinja",
         "executor": {
-            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/common:2.3.1"}
+            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/common:3.0.0"}
         },
     }
     data_name = '{{ reads|sample_name|default("?") }}'