+++ resolwe_bio/processes/reads_processing/alignmentsieve.py
@@ -28,7 +28,7 @@ class AlignmentSieve(Process):
     slug = "alignmentsieve"
     name = "alignmentSieve"
     process_type = "data:alignment:bam:sieve"
-    version = "1.2.1"
+    version = "1.3.0"
     category = "Alignment Filtering"
     data_name = 'Sieved BAM ({{ alignment|sample_name|default("?") }})'
     scheduling_class = SchedulingClass.BATCH
@@ -39,7 +39,7 @@ class AlignmentSieve(Process):
     requirements = {
         "expression-engine": "jinja",
         "executor": {
-            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/common:2.6.0"}
+            "docker": {"image": "public.ecr.aws/s4q6j6e8/resolwebio/common:3.0.0"}
         },
         "resources": {
             "cores": 10,